attributes (#2)
* add a tag with attribute href to index.htm

* add img tags with relative source attributes to index.htm

- If the URL begins without a slash, it will be relative to the current page. Example: src="img.png".
- If the URL begins with a slash, it will be relative to the domain. Example: src="/img.png".

* add width and height attributes to new img tags of index.htm

* add alt attributes to img tags of index.htm

The required alt attribute for the <img> tag specifies an alternate text for an image, if the image for some reason cannot be displayed. This can be due to a slow connection, or an error in the src attribute, or if the user uses a screen reader.

* add style attribute to first p tag of index.htm

* add lang attribute html tag

You should always include the lang attribute inside the <html> tag, to declare the language of the Web page. This is meant to assist search engines and browsers.

* extend lang attribute of html tag by country code

Country codes can also be added to the language code in the lang attribute. So, the first two characters define the language of the HTML page, and the last two characters define the country.

* add title attribute to second p tag

The title attribute defines some extra information about an element.

The value of the title attribute will be displayed as a tooltip when you mouse over the element

* add second title attribute to third p tag with usage of single quotes around attributes value

Double quotes around attribute values are the most common in HTML, but single quotes can also be used.

In some situations, when the attribute value itself contains double quotes, it is necessary to use single quotes