add two paragraphs seperated by horizontal rules to index.htm
The <hr> tag defines a thematic break in an HTML page, and is most often displayed as a horizontal rule.

The <hr> element is used to separate content (or define a change) in an HTML page

The <hr> tag is an empty tag, which means that it has no end tag.