links (#11)
* add basic html page for link examples

* add first a tag example

By default, links will appear as follows in all browsers:
- An unvisited link is underlined and blue
- A visited link is underlined and purple
- An active link is underlined and red

* add examples for usage of target attribute
- _self - Default. Opens the document in the same window/tab as it was clicked
- _blank - Opens the document in a new window or tab
- _parent - Opens the document in the parent frame
- _top - Opens the document in the full body of the window

* add examples absolute and relative urls

* add example for a tag with image

* add example for a tag mailto href attribute

* add example for button as link

To use an HTML button as a link, you have to add some JavaScript code. JavaScript allows you to specify what happens at certain events, such as a click of a button.

* add example for a tag with title attribute

The title attribute specifies extra information about an element. The information is most often shown as a tooltip text when the mouse moves over the element.

* add styling for links

Here, an unvisited link will be green with no underline. A visited link will be pink with no underline. An active link will be yellow and underlined. In addition, when mousing over a link (a:hover) it will become red and underlined.

* add lorem ipsum text to fill page

* add a tags with bookmarks in href attribute