images (#12)
* add images.htm file for images exampels

* add exmaple for image tag

The HTML `<img>` tag is used to embed an image in a web page.

Images are not technically inserted into a web page; images are linked to web pages. The `<img>` tag creates a holding space for the referenced image.

The `<img>` tag is empty, it contains attributes only, and does not have a closing tag.

The `<img>` tag has two required attributes:

- src - Specifies the path to the image
- alt - Specifies an alternate text for the image

* add example for image tag with specific height and width

* add example for image with external source

* add example for image with float styling

Use the CSS float property to let the image float to the right or to the left of a text.