tabels (#18)
* add table.htm file for table examples

* add basic table example

Each table cell is defined by a <td> and a </td> tag.

td stands for table data.

Everything between <td> and </td> are the content of the table cell.

* add second table with only one row

Each table row starts with a <tr> and ends with a </tr> tag.

tr stands for table row.

* add second row to second table

You can have as many rows as you like in a table; just make sure that the number of cells are the same in each row.

Note: There are times when a row can have less or more cells than another. You will learn about that in a later chapter.

* add table header to second table

Sometimes you want your cells to be table header cells. In those cases use the <th> tag instead of the <td> tag.

th stands for table header.

By default, the text in <th> elements are bold and centered, but you can change that with CSS.

* add collapsed border to table

To avoid having double borders like in the example above, set the CSS border-collapse property to collapse.

This will make the borders collapse into a single border

* add invisible border table example

If you set a background color of each cell, and give the border a white color (the same as the document background), you get the impression of an invisible border.

* add rounded border table example

- With the border-radius property, the borders get rounded corners.

- Skipping the border around the table makes the items appear seperated

* add border style table examples for:
- dotted
- dashed
- double
- groove
- ridge
- inset
- outset

* extract border styling from shorthand to long format and add border-color

* add empty line to end of table.htm

* add table example with width property

Note: Using a percentage as the size unit for a width means how wide will this element be compared to its parent element, which in this case is the <body> element.

* add width to single column

To set the size of a specific column, add the style attribute on a <th> or <td> element.

* add height to single row

To set the height of a specific row, add the style attribute on a table row element.
(added border styling to make height visible)

* add table example with vertical header

HTML tables can have headers for each column or row, or for many columns/rows. Table headers are defined with th elements. Each th element represents a table cell.

To use the first column as table headers, define the first cell in each row as a <th> element

* align table header left

* add example for colspan

You can have a header that spans over two or more columns. To do this, use the colspan attribute on the <th> element.

* add example for caption to table

You can add a caption that serves as a heading for the entire table. To add a caption to a table, use the <caption> tag.

Note: The <caption> tag should be inserted immediately after the <table> tag.

* add table with padding

Cell padding is the space between the cell edges and the cell content.
By default the padding is set to 0.
To add padding on table cells, use the CSS padding property

To add padding only above the content, use the padding-top property.

And the others sides with the padding-bottom, padding-left, and padding-right properties:
th, td {
  padding-top: 10px;
  padding-bottom: 20px;
  padding-left: 30px;
  padding-right: 40px;
}

* add border-spacing to table

Cell spacing is the space between each cell.

By default, the space is set to 2 pixels.

To change the space between table cells, use the CSS border-spacing property on the table element

* add table examples for colspan and rowspan

HTML tables can have cells that span over multiple rows and/or columns.

 -To make a cell span over multiple columns, use the colspan attribute
 - To make a cell span over multiple rows, use the rowspan attribute

 NOTE: The value of the colspan/rowspan attribute represents the number of columns/rows to span

* add table examples for zebra effect

If you add a background color on every other table row, you will get a nice zebra stripes effect. Do the same for the td tad and if will have the vertical zebra effect

Note: If you use (odd) instead of (even), the styling will occur on row 1,3,5 etc. instead of 2,4,6 etc.

* add table examples combined vertical and horizontal zebra stripes

You can combine the styling from the two examples above and you will have stripes on every other row and every other column.

If you use a transparent color you will get an overlapping effect.

* FIX add example for table with horizontal dividers

If you specify borders only at the bottom of each table row, you will have a table with horizontal dividers.

Add the border-bottom property to all tr elements to get horizontal dividers

* add example for hoverable table

Use the :hover selector on tr to highlight table rows on mouse over

* add example table with styled colgroups

There is only a very limited selection of CSS properties that are allowed to be used in the colgroup:

width property
visibility property
background properties
border properties

All other CSS properties will have no effect on your tables

Note: The <colgroup> tag must be a child of a <table> element and should be placed before any other table elements, like <thead>, <tr>, <td> etc., but after the <caption> element, if present.

* remove duplicate semicolon