block-inline (#23)
* add block-inline.htm for examples

* add examples for block elements

* add examples for inline element inside p tag

Note: An inline element cannot contain a block-level element!

* add div tag with styling

The <div> element is often used as a container for other HTML elements. The <div> element has no required attributes, but style, class and id are common. When used together with CSS, the <div> element can be used to style blocks of content

* add p tag with styled span tags

The <span> element is an inline container used to mark up a part of a text, or a part of a document. The <span> element has no required attributes, but style, class and id are common. When used together with CSS, the <span> element can be used to style parts of the text