add example for usage of multiple classes
HTML elements can belong to more than one class. To define multiple classes, separate the class names with a space, e.g. <div class="city main">. The element will be styled according to all the classes specified.

In this example, the first <h2> element belongs to both the city class and also to the main class, and will get the CSS styles from both of the classes