classes (#24)
* add classes.htm for examples

* add div tags with styling assigned using class

The class attribute is often used to point to a class name in a style sheet. It can also be used by a JavaScript to access and manipulate elements with the specific class name.

In the example we have three <div> elements with a class attribute with the value of "city". All of the three <div> elements will be styled equally according to the .city style definition in the head section

* add span tags with styling assigned using class

In this example we have two <span> elements with a class attribute with the value of "note". Both <span> elements will be styled equally according to the .note style definition in the head section

* add example for usage of multiple classes

HTML elements can belong to more than one class. To define multiple classes, separate the class names with a space, e.g. <div class="city main">. The element will be styled according to all the classes specified.

In this example, the first <h2> element belongs to both the city class and also to the main class, and will get the CSS styles from both of the classes

* add example of class usage within two different tags

Different HTML elements can point to the same class name. In the example, both <h2> and <p> point to the "city" class and will share the same style

* add example of class usage to access within JavaScript

The class name can also be used by JavaScript to perform certain tasks for specific elements. JavaScript can access elements with a specific class name with the getElementsByClassName() method