head-tag (#29)
* add head.htm for examples

* change title of head.htm

The <title> element defines the title of the document. The title must be text-only, and it is shown in the browser's title bar or in the page's tab.
The <title> element is required in HTML documents!
The content of a page title is very important for search engine optimization (SEO)! The page title is used by search engine algorithms to decide the order when listing pages in search results.

The <title> element:
- defines a title in the browser toolbar
- provides a title for the page when it is added to favorites
- displays a title for the page in search engine-results

* add style tag to head tag

The <style> element is used to define style information for a single HTML page

* add link tag to head tag

The <link> element defines the relationship between the current document and an external resource. The <link> tag is most often used to link to external style sheets

* add meta tag for character set used to head tag

The <meta> element is typically used to specify the character set, page description, keywords, author of the document, and viewport settings.
The metadata will not be displayed on the page, but is used by browsers (how to display content or reload page), by search engines (keywords), and other web services.

* add meta tag with keywords for search engines

* add meta tag for description of webpage

* add meta tag for author of webpage

* add meta tag for browsers to refresh page every 30 seconds

* add meta tag to set the viewport, to make your website look good on all devices

This gives the browser instructions on how to control the page's dimensions and scaling. The width=device-width part sets the width of the page to follow the screen-width of the device (which will vary depending on the device).
The initial-scale=1.0 part sets the initial zoom level when the page is first loaded by the browser.

* add script tag to head tag

The <script> element is used to define client-side JavaScripts.
The example JavaScript writes "Hello JavaScript!" into an HTML element with id="demo"

* add base tag to head tag

The <base> element specifies the base URL and/or target for all relative URLs in a page.
The <base> tag must have either an href or a target attribute present, or both.
There can only be one single <base> element in a document!