responsive (#31)
* add responsive.htm for examples

* add viewport meta tag to head

This will set the viewport of the page, which will give the browser instructions on how to control the page's dimensions and scaling.

* add image with width given in percent

If the CSS width property is set to 100%, the image will be responsive and scale up and down.
Notice: The image can be scaled up to be larger than its original size. A better solution, in many cases, will be to use the max-width property instead.

* add image with max-width property

* add picture tag to display images based on device width

The HTML <picture> element allows you to define different images for different browser window sizes.

* add h1 tag with font-size property of viewport width

The text size can be set with a "vw" unit, which means the "viewport width". That way the text size will follow the size of the browser window.

Viewport is the browser window size. 1vw = 1% of viewport width. If the viewport is 50cm wide, 1vw is 0.5cm.

* add example for media query

In addition to resize text and images, it is also common to use media queries in responsive web pages.
With media queries you can define completely different styles for different browser sizes.
In this example the three div elements will display horizontally on large screens and stack vertically on small screens.

* add full response example

* add full response example using W3.CSS

* add full response example using Bootstrap