computercode (#32)
* add computercode.htm for examples

* add code tag example

The HTML <code> element  is used to define a piece of computer code. The content inside is displayed in the browser's default monospace font.

* add kdb tag example

The HTML <kbd> element is used to define keyboard input. The content inside is displayed in the browser's default monospace font.

* add samp tag example

The HTML <samp> element is used to define sample output from a computer program. The content inside is displayed in the browser's default monospace font.

* add code tag example surrounded by pre tag

The <code> element does not preserve extra whitespace and line-breaks.
To fix this, you can put the <code> element inside a <pre> element

* add var tag example

The HTML <var> element  is used to define a variable in programming or in a mathematical expression. The content inside is typically displayed in italic.