add example for section tags
The <section> element defines a section in a document.

According to W3C's HTML documentation: "A section is a thematic grouping of content, typically with a heading."

Examples of where a <section> element can be used:

    Chapters
    Introduction
    News items
    Contact information

A web page could normally be split into sections for introduction, content, and contact information.