Never Skip the <title> Element
The <title> element is required in HTML.

The contents of a page title is very important for search engine optimization (SEO)! The page title is used by search engine algorithms to decide the order when listing pages in search results.

The <title> element:

    defines a title in the browser toolbar
    provides a title for the page when it is added to favorites
    displays a title for the page in search-engine results

So, try to make the title as accurate and meaningful as possible