style-guide (#34)
* add style-guide.htm

* Add document type declaration

Always declare the document type as the first line in your document

* Use Lowercase Element Names

HTML allows mixing uppercase and lowercase letters in element names.
However, we recommend using lowercase element names, because:
    Mixing uppercase and lowercase names looks bad
    Developers normally use lowercase names
    Lowercase looks cleaner
    Lowercase is easier to write

* Close All HTML Elements

In HTML, you do not have to close all elements (for example the <p> element).
However, we strongly recommend closing all HTML elements, like this

* add example for closed image tag

The example provided ist compliant with HTML5 and XHTML

* Use Lowercase Attribute Names

HTML allows mixing uppercase and lowercase letters in attribute names.
However, we recommend using lowercase attribute names, because:

    Mixing uppercase and lowercase names looks bad
    Developers normally use lowercase names
    Lowercase looks cleaner
    Lowercase is easier to write

* Always Quote Attribute Values

HTML allows attribute values without quotes.

However, we recommend quoting attribute values, because:

    Developers normally quote attribute values
    Quoted values are easier to read
    You MUST use quotes if the value contains spaces

* Always Specify alt, width, and height for Images

Always specify the alt attribute for images. This attribute is important if the image for some reason cannot be displayed.

Also, always define the width and height of images. This reduces flickering, because the browser can reserve space for the image before loading.

* Spaces and Equal Signs

HTML allows spaces around equal signs. But space-less is easier to read and groups entities better together.

* Avoid Long Code Lines

When using an HTML editor, it is NOT convenient to scroll right and left to read the HTML code.

Try to avoid too long code lines.

* Blank Lines and Indentation

Do not add blank lines, spaces, or indentations without a reason.

For readability, add blank lines to separate large or logical code blocks.

For readability, add two spaces of indentation. Do not use the tab key.

* Add example for indentation

In this example the default of the IDE was used and because of that the indentation is set to 4 spaces for every level

* Never Skip the <title> Element

The <title> element is required in HTML.

The contents of a page title is very important for search engine optimization (SEO)! The page title is used by search engine algorithms to decide the order when listing pages in search results.

The <title> element:

    defines a title in the browser toolbar
    provides a title for the page when it is added to favorites
    displays a title for the page in search-engine results

So, try to make the title as accurate and meaningful as possible

* Closing empty element

In HTML, it is optional to close empty elements. If you expect XML/XHTML software to access your page, keep the closing slash (/), because it is required in XML and XHTML.

* Add the lang Attribute

You should always include the lang attribute inside the <html> tag, to declare the language of the Web page. This is meant to assist search engines and browsers.

* Meta Data

To ensure proper interpretation and correct search engine indexing, both the language and the character encoding <meta charset="charset"> should be defined as early as possible in an HTML document

* Setting the Viewport

The viewport is the user's visible area of a web page. It varies with the device - it will be smaller on a mobile phone than on a computer screen.

This gives the browser instructions on how to control the page's dimensions and scaling.

The width=device-width part sets the width of the page to follow the screen-width of the device (which will vary depending on the device).

The initial-scale=1.0 part sets the initial zoom level when the page is first loaded by the browser.

* HTML Comments

Short comments should be written on one line.
Comments that spans more than one line should be formatted so that they are easier to observe.

* Using Style Sheets

Short CSS rules can be written compressed.

Long CSS rules should be written over multiple lines.

    Place the opening bracket on the same line as the selector
    Use one space before the opening bracket
    Use semicolon after each property-value pair, including the last
    Only use quotes around values if the value contains spaces
    Place the closing bracket on a new line, without leading spaces

* Loading JavaScript in HTML

Use simple syntax for loading external scripts (the type attribute is not necessary).

* Accessing HTML Elements with JavaScript

Using "untidy" HTML code can result in JavaScript errors.
These two JavaScript statements will produce different results

* Use Lower Case File Names

Some web servers (Apache, Unix) are case-sensitive about file names: "london.jpg" cannot be accessed as "London.jpg".

Other web servers (Microsoft, IIS) are not case-sensitive: "london.jpg" can be accessed as "London.jpg".

If you use a mix of uppercase and lowercase, you have to be aware of this.

If you move from a case-insensitive to a case-sensitive server, even small errors will break your web!

To avoid these problems, always use lowercase file names!

* File Extensions

HTML files should have a .html extension (.htm is allowed).

CSS files should have a .css extension.

JavaScript files should have a .js extension.

* Differences Between .htm and .html?

There is no difference between the .htm and .html file extensions!

Both will be treated as HTML by any web browser and web server.

* Default Filenames

When a URL does not specify a filename at the end (like "https://www.w3schools.com/"), the server just adds a default filename, such as "index.html", "index.htm", "default.html", or "default.htm".

If your server is configured only with "index.html" as the default filename, your file must be named "index.html", and not "default.html".

However, servers can be configured with more than one default filename; usually you can set up as many default filenames as you want.